멀티해상도 이슈 해결 (안드로이드, 아이폰)

+++ puzzlezip/Classes/Common/CustomMath.cpp
@@ -7,14 +7,15 @@
 //
 
 #include "CustomMath.h"
+#include <math.h>
 
 int getPtDistance( Vec2 pt1, Vec2 pt2 )
 {
     return getPtDistance(pt1.x, pt1.y, pt2.x, pt2.y);
 }
 
-int getPtDistance( int ptx, int pty, int pt2x, int pt2y )
+int getPtDistance( float ptx, float pty, float pt2x, float pt2y )
 {
-    return sqrt( fabsf((ptx - pt2x))*fabsf((ptx - pt2x)) + fabsf((pty - pt2y))*fabsf((pty - pt2y)) );
+    return sqrt( std::fabsf((ptx - pt2x))*std::fabsf((ptx - pt2x)) + std::fabsf((pty - pt2y))*std::fabsf((pty - pt2y)) );
 }
